+++ docs/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Changelog for v2.2.2
 - Read CPG file attached to the shapefile- 
 - Declare the schema for the spatial_ref_sys table
+- Add SnapToGrid function
 - Write empty geometry with the shapefile driver